feat(agent): add export method
